requirements_npu.txt
colossalai
accelerate
diffusers
ftfy
gdown
mmengine
pandas
pre-commit
pyarrow
av
tensorboard
timm
tqdm
transformers
wandb
rotary_embedding_torch
pandarallel
scipy
decorator
attrs
